backend/requirements.txt
fastapi
uvicorn
python-multipart
whisper
ffmpeg-python
pydantic